+++ Dfe.Academies.Academisation.Service/Queries/SignificantChange/GetSignificantProjectsQueryHandler.cs
@@ -2,6 +2,7 @@
 using Dfe.Academies.Academisation.IService.ServiceModels.Legacy.ProjectAggregate;
 using Dfe.Academies.Academisation.IService.ServiceModels.SignificantChange;
 using Dfe.Academies.Academisation.Service.Factories;
+using Dfe.Academies.Academisation.Service.Mappers.SignificantChange;
 using MediatR;
 
 namespace Dfe.Academies.Academisation.Service.Queries.SignificantChange
@@ -12,30 +13,15 @@ namespace Dfe.Academies.Academisation.Service.Queries.SignificantChange
 
 		public async Task<PagedDataResponse<SignificantChangeProjectSearchResponse>> Handle(GetSignificantProjectsQuery query, CancellationToken cancellationToken)
 		{
-		var (projects, totalCount) = await _significantChangeProjectRepository.SearchSignificantProjects(query.Page, query.Count, cancellationToken);
+			var (projects, totalCount) = await _significantChangeProjectRepository.SearchSignificantProjects(query.Page, query.Count, cancellationToken);
 
-		var routeValues = new Dictionary<string, object?>();
-		var pageResponse = PagingResponseFactory.Create("significant-change/significant-change-projects", query.Page, query.Count, totalCount, routeValues);
-	
-		var data = projects.Select(p => new SignificantChangeProjectSearchResponse
-		{
-			Id = p.Id,
-			Urn = p.Urn,
-			SchoolName = p.SchoolName,
-			Tier = p.Tier,
-			TrustName = p.TrustName,
-			TrustUkprn = p.TrustUkprn,
-			AssignedUser = p.AssignedUserId != null ? new User(
-				p.AssignedUserId.Value,
-				p.AssignedUserFullName ?? string.Empty,
-				p.AssignedUserEmailAddress ?? string.Empty
-			) : null,
-			TypeOfSignificantChange = p.TypeOfSignificantChange,
-			Status = p.Status.ToString()
-		});
+			var routeValues = new Dictionary<string, object?>();
+			var pageResponse = PagingResponseFactory.Create("significant-change/significant-change-projects", query.Page, query.Count, totalCount, routeValues);
 
-		return new PagedDataResponse<SignificantChangeProjectSearchResponse>(data,
-			pageResponse);
+			var data = projects.Select(p => p.MapToServiceModel());
+
+			return new PagedDataResponse<SignificantChangeProjectSearchResponse>(data,
+				pageResponse);
 		}
 
 	}